+++ CHANGELOG
@@ -1,5 +1,10 @@
 All notable changes to this project will be documented in this file.
 
+Uncommitted changes
+    * Improved setup file
+    * Simplification of nesting and coding style throughout
+    * Cleaned up all code files and made style more consistent using pep8 style 
+
 2019-09-11 - version 0.1.3
     * Updated Readme and setup.py
     * Added Changelog
