--- CONTRIBUTING.md
@@ -1,10 +1,0 @@
-Coding Style
-------
-For consistency, please ensure that all files are formatted by `yapf` using the pep8 style.  Please run the following commands:
-```shell
-pip install yapf
-```
-Please reformat all source files in the src `./movekit/` folder:
-```shell
-yapf -i -r movekit/
-```
