+++ movekit/__init__.py
@@ -1,1 +1,5 @@
-__path__ = __import__('pkgutil').extend_path(__path__, __name__)
+#! /usr/bin/env python
+
+from .feature_extraction import *
+from .preprocess import *
+from .utils import *
