+++ movekit/__version__.py
@@ -1,3 +1,5 @@
+#! /usr/bin/env python
+
 major = 0
 minor = 1
 micro = 4
