--- movekit/feature_extraction/__init__.py
@@ -1,5 +1,0 @@
-from .absolute import *
-from .group import *
-from .comparative import *
-from .absolute_agg import *
-name = 'feature_extraction'
