+++ movekit/tests/data/missing-col.csv
@@ -1,0 +1,51 @@
+time,x,y
+1,405.29,417.76
+1,369.99,428.78
+1,390.33,405.89
+1,445.15,411.94
+1,366.06,451.76
+2,405.31,417.37
+2,370.01,428.82
+2,390.25,405.89
+2,445.48,412.26
+2,365.86,451.76
+3,405.31,417.07
+3,370.01,428.85
+3,390.17,405.88
+3,445.77,412.61
+3,365.7,451.76
+4,405.3,416.86
+4,370.01,428.86
+4,390.07,405.88
+4,446.03,413
+4,365.57,451.76
+5,405.29,416.71
+5,369.99,428.86
+5,389.98,405.87
+5,446.24,413.42
+5,365.47,451.76
+6,405.27,416.61
+6,369.98,428.84
+6,389.88,405.87
+6,446.42,413.86
+6,365.39,451.76
+7,405.27,416.54
+7,369.97,428.82
+7,389.79,405.87
+7,446.54,414.34
+7,365.35,451.76
+8,405.27,416.49
+8,369.97,428.8
+8,389.7,405.88
+8,446.62,414.85
+8,365.33,451.76
+9,405.31,416.37
+9,369.98,428.82
+9,389.68,405.9
+9,446.63,415.42
+9,365.34,451.76
+10,405.38,416.27
+10,370.08,428.81
+10,389.62,405.93
+10,446.55,415.92
+10,365.36,451.81
