+++ movekit/tests/data/missing-header.csv
@@ -1,0 +1,50 @@
+1,312,405.29,417.76
+1,511,369.99,428.78
+1,607,390.33,405.89
+1,811,445.15,411.94
+1,905,366.06,451.76
+2,312,405.31,417.37
+2,511,370.01,428.82
+2,607,390.25,405.89
+2,811,445.48,412.26
+2,905,365.86,451.76
+3,312,405.31,417.07
+3,511,370.01,428.85
+3,607,390.17,405.88
+3,811,445.77,412.61
+3,905,365.7,451.76
+4,312,405.3,416.86
+4,511,370.01,428.86
+4,607,390.07,405.88
+4,811,446.03,413
+4,905,365.57,451.76
+5,312,405.29,416.71
+5,511,369.99,428.86
+5,607,389.98,405.87
+5,811,446.24,413.42
+5,905,365.47,451.76
+6,312,405.27,416.61
+6,511,369.98,428.84
+6,607,389.88,405.87
+6,811,446.42,413.86
+6,905,365.39,451.76
+7,312,405.27,416.54
+7,511,369.97,428.82
+7,607,389.79,405.87
+7,811,446.54,414.34
+7,905,365.35,451.76
+8,312,405.27,416.49
+8,511,369.97,428.8
+8,607,389.7,405.88
+8,811,446.62,414.85
+8,905,365.33,451.76
+9,312,405.31,416.37
+9,511,369.98,428.82
+9,607,389.68,405.9
+9,811,446.63,415.42
+9,905,365.34,451.76
+10,312,405.38,416.27
+10,511,370.08,428.81
+10,607,389.62,405.93
+10,811,446.55,415.92
+10,905,365.36,451.81
